requirements.txt
aomaker>=2.1.0; python_version >= "3.8"
faker==13.12.0
python-jenkins==1.7.0
tenacity==8.0.1
openpyxl==3.0.10
psutil==5.9.1
netaddr==0.8.0
psycopg2-binary==2.9.3
redis==4.2.2
paramiko==2.11.0
pytest-rerunfailures==10.2
pytest-repeat==0.9.1
IPy==1.1
qingcloud-sdk==1.2.15
pytest-dependency==0.5.1
filelock==3.10.6